fd: Introduce Evaluable mixin
